Move groups to groups package
